+++ ui/components/providers/radio-group-provider.tsx
@@ -12,8 +12,10 @@ import {
   AzureProviderBadge,
   GCPProviderBadge,
   GitHubProviderBadge,
+  IacProviderBadge,
   KS8ProviderBadge,
   M365ProviderBadge,
+  OracleCloudProviderBadge,
 } from "../icons/providers-badge";
 import { CustomRadio } from "../ui/custom";
 import { FormMessage } from "../ui/form";
@@ -76,6 +78,21 @@ export const RadioGroupProvider: React.FC<RadioGroupProviderProps> = ({
                 <div className="flex items-center">
                   <GitHubProviderBadge size={26} />
                   <span className="ml-2">GitHub</span>
+                </div>
+              </CustomRadio>
+              <CustomRadio description="Infrastructure as Code" value="iac">
+                <div className="flex items-center">
+                  <IacProviderBadge size={26} />
+                  <span className="ml-2">Infrastructure as Code</span>
+                </div>
+              </CustomRadio>
+              <CustomRadio
+                description="Oracle Cloud Infrastructure"
+                value="oraclecloud"
+              >
+                <div className="flex items-center">
+                  <OracleCloudProviderBadge size={26} />
+                  <span className="ml-2">Oracle Cloud Infrastructure</span>
                 </div>
               </CustomRadio>
             </div>